Docummentation for the MIMO classes.
The code in the encode and decode methods in the Blast and Alamouti classes in the mimo module where moved to the new _encode and _decode methods. The encode and decode methods call _encode and _decode and then perform the power division/compensation. This will alow better code reuse when the hybrid MIMO schemes are implemented.

Implementation of tests for the Alamouiti class.

git-svn-id: http://py-physim.googlecode.com/svn/trunk@70 52f46f9c-28e8-69bd-04f8-aaf5e2abe915