Modified the is_point_inside_shape method of the Shape class.
The shape class used the matplotlib.nxutils.pnpoly function to test if a
point is inside of the shape. However, this function is deprecated since
matplotlib 1.2 and now it needs to create a matplotlib.path.Path object and
use its contains_point method to test of a point is inside the (Shape's)
path.




git-svn-id: http://py-physim.googlecode.com/svn/trunk@232 52f46f9c-28e8-69bd-04f8-aaf5e2abe915