Implemented the saving of partial results feature in the simulate_in_parallel method of the SimulationRunner class.
Note that there is a limitation. The partial results files will be saved in
the folder where the IPython cluster was started. However, the deletion of
the files is not performed in by the IPython engines, but by the main
python program. Therefore, unless the Ipython engines are running in the
same folder where the main python program will be run the partial result
files won't be automatically deleted after the simulation is finished.



git-svn-id: http://py-physim.googlecode.com/svn/trunk@376 52f46f9c-28e8-69bd-04f8-aaf5e2abe915