Test for a numpy bug
When numpy is manually installed, but without lapack, it will use its
own internal SVD implementation. This implementation has a bug for
complex matrices with more columns then rows, where the returned right
singular vectors for the 0 valued singular vectors will be all zeros. We
test this here to make sure the installed version of numpy was compiled
with lapack to avoid this bug.