Fix simulate_comp.py after change in the channel
At some commit in the past the corrupt data methods, as well as the
methods to calculate the SINR do not receive the noise variance
anymore. Instead the noise variance property of the channel class must
be set before calling those methods.

In this commit simulate_comp.py is changed to be consistent with this
past change in the channel classes.