New SkipThisOne exception to skip _run_simulation
Add the new SkipThisOne exception. If this exception is raised in the
_run_simulation implementation in a subclass of SimulationRunner then
that repetition will be skipped and the new 'num_skipped_reps' result
will be increased to log that event.

The merge_all_results method of the SimulationResults class will allow
merging two SimulationResults objects even if one of then does not have
the 'num_skipped_reps' result.