Raise SkipThisOne exception on IA algorithm error
If the MMSE IA algorithm raises a RuntimeError or a LinAlgError
exception and thus don't find a solution, then the SimulationRunner will
raise a SkipThisOne exception to skip the current repetition. Hopefully
in the next one (with new channel and precoder initialization) the IA
algorithm will find a solution.

This change makes the simulator robust ill conditioned conditions,
which will simply be skipped.