Add a more parameters to greedy simulator
Particularly, add the "stream_sel_method" parameter to select the stream
selection algorithm (either greedy or brute force) instead of simulating
both cases in the same 'simulation'. This allows taking advantage of the
parallel capabilities of the SimulationRunenr class.

The config spec was also updated/improved to reflect the changes in the
simulation parameters.