Handle pathloss for small distances
Add the "handle_small_distances_bool" to the PathLossBase class. If it
is set to False then an exception is raised if the path loss is
calculated for a distance so small that it would be negative (in dB).

If handle_small_distances_bool is set to True then in case the path loss
is negative (in dB) it is assumed to be equal to 0dB.