Add ipcluster2 and ipcluster3 scripts
The ipcluster script is installed when you install the ipyparallel
package (using pip). Depending if you run pip with python2 or python3 it
will install the corresponding version, but you can't have both.

Here we add the same script only changing the shebang to allow testing
with both python versions. That is, in the tests folder you should run
`./ipcluster2 start --profile=tests`
if you want to run the pyphysim tests with python2 or
`./ipcluster3 start --profile=tests`
if you want to run the pyphysim tests with python3.