Reorganize some simulators and update their code
Reorganize some simulators such as the MIMO simulators, the simulators
for PSK/QAM/etc in AWGN channel, and the IA simulators.

These simulators are put in their own folder.

Also, the code was updated to take into account changes in the
libraries (both external libraries, such as ipython related, as well as
the internal libraries such as the simulationresultsplotter which was
removed).