Change parse function in configobjvalidation.py to return lists instead of numpy arrays
The reason if that the return is used at some point to compare with another
value by configobj validation inside an if statement. Since numpy arrays perform
comparison element-by-element it will raise an exception if you try to use the
comparison in an if statement (as configobj library does). Lists, on the other
hand, compare the whole list and are compatible with configobj validation.